Readme updates and FW19111 last mods

+++ Sopranino/2GW/FW19111/esp32_arduino_19X_19111/esp32_arduino_19X_19111.ino
@@ -108,7 +108,7 @@ struct Tstick {
   int32_t firmware;
   char oscIP[17];
   int32_t oscPORT;
-  byte FSRoffset;
+  float FSRoffset;
   byte touchMask[2];
   float abias[3];
   float mbias[3];
